refactor(graph) graph into separate component
